heurist-v0.001/agents-service/requirements.txt
fastapi
uvicorn
crewai
requests
python-dotenv
